+++ src/suou/__init__.py
@@ -35,20 +35,19 @@ from .strtools import PrefixIdentifier
 from .validators import matches, not_less_than, not_greater_than, yesno
 from .redact import redact_url_password
 from .http import WantsContentType
-from .color import chalk, WebColor
+from .color import OKLabColor, chalk, WebColor, RGBColor, SRGBColor, XYZColor, OKLabColor
 from .mat import Matrix
 
-__version__ = "0.12.0a4"
+__version__ = "0.12.0a5"
 
 __all__ = (
     'ConfigOptions', 'ConfigParserConfigSource', 'ConfigSource', 'ConfigValue',
     'DictConfigSource', 'EnvConfigSource', 'I18n', 'Incomplete', 'JsonI18n',
-    'Matrix',
-    'MissingConfigError', 'MissingConfigWarning', 'PrefixIdentifier',
-    'RGBColor', 'SRGBColor',
+    'Matrix', 'MissingConfigError', 'MissingConfigWarning', 'OKLabColor', 
+    'PrefixIdentifier', 'RGBColor', 'SRGBColor',
     'Siq', 'SiqCache', 'SiqGen', 'SiqType', 'Snowflake', 'SnowflakeGen',
     'StringCase', 'TimedDict', 'TomlI18n', 'UserSigner', 'Wanted', 'WantsContentType',
-    'WebColor',
+    'WebColor', 'XYZColor',
     'addattr', 'additem', 'age_and_days', 'alru_cache', 'b2048decode', 'b2048encode',
     'b32ldecode', 'b32lencode', 'b64encode', 'b64decode', 'cb32encode',
     'cb32decode', 'chalk', 'count_ones', 'dei_args', 'deprecated',